Move test to right place
