+++ sca/bom/buildinfo/technologies/common.go
@@ -18,7 +18,6 @@ import (
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
 	"github.com/jfrog/jfrog-client-go/utils/log"
 	"github.com/jfrog/jfrog-client-go/xray/services"
-	xrayUtils "github.com/jfrog/jfrog-client-go/xray/services/utils"
 	xscservices "github.com/jfrog/jfrog-client-go/xsc/services"
 )
 
@@ -98,94 +97,6 @@ func LogExecutableVersion(executable string) {
 	}
 	version := strings.TrimSpace(string(verBytes))
 	log.Debug(fmt.Sprintf("Used %q version: %s", executable, version))
-}
-
-// BuildImpactPathsForScanResponse builds the full impact paths for each vulnerability found in the scanResult argument, using the dependencyTrees argument.
-// Returns the updated services.ScanResponse slice.
-func BuildImpactPathsForScanResponse(scanResult []services.ScanResponse, dependencyTree []*xrayUtils.GraphNode) []services.ScanResponse {
-	for _, result := range scanResult {
-		if len(result.Vulnerabilities) > 0 {
-			buildVulnerabilitiesImpactPaths(result.Vulnerabilities, dependencyTree)
-		}
-		if len(result.Violations) > 0 {
-			buildViolationsImpactPaths(result.Violations, dependencyTree)
-		}
-		if len(result.Licenses) > 0 {
-			buildLicensesImpactPaths(result.Licenses, dependencyTree)
-		}
-	}
-	return scanResult
-}
-
-// Initialize a map of issues empty impact paths
-func fillIssuesMapWithEmptyImpactPaths(issuesImpactPathsMap map[string][][]services.ImpactPathNode, components map[string]services.Component) {
-	for dependencyName := range components {
-		issuesImpactPathsMap[dependencyName] = [][]services.ImpactPathNode{}
-	}
-}
-
-// Set the impact paths for each issue in the map
-func buildImpactPaths(issuesImpactPathsMap map[string][][]services.ImpactPathNode, dependencyTrees []*xrayUtils.GraphNode) {
-	for _, dependency := range dependencyTrees {
-		setPathsForIssues(dependency, issuesImpactPathsMap, []services.ImpactPathNode{})
-	}
-}
-
-func buildVulnerabilitiesImpactPaths(vulnerabilities []services.Vulnerability, dependencyTrees []*xrayUtils.GraphNode) {
-	issuesMap := make(map[string][][]services.ImpactPathNode)
-	for _, vulnerability := range vulnerabilities {
-		fillIssuesMapWithEmptyImpactPaths(issuesMap, vulnerability.Components)
-	}
-	buildImpactPaths(issuesMap, dependencyTrees)
-	for i := range vulnerabilities {
-		updateComponentsWithImpactPaths(vulnerabilities[i].Components, issuesMap)
-	}
-}
-
-func buildViolationsImpactPaths(violations []services.Violation, dependencyTrees []*xrayUtils.GraphNode) {
-	issuesMap := make(map[string][][]services.ImpactPathNode)
-	for _, violation := range violations {
-		fillIssuesMapWithEmptyImpactPaths(issuesMap, violation.Components)
-	}
-	buildImpactPaths(issuesMap, dependencyTrees)
-	for i := range violations {
-		updateComponentsWithImpactPaths(violations[i].Components, issuesMap)
-	}
-}
-
-func buildLicensesImpactPaths(licenses []services.License, dependencyTrees []*xrayUtils.GraphNode) {
-	issuesMap := make(map[string][][]services.ImpactPathNode)
-	for _, license := range licenses {
-		fillIssuesMapWithEmptyImpactPaths(issuesMap, license.Components)
-	}
-	buildImpactPaths(issuesMap, dependencyTrees)
-	for i := range licenses {
-		updateComponentsWithImpactPaths(licenses[i].Components, issuesMap)
-	}
-}
-
-func updateComponentsWithImpactPaths(components map[string]services.Component, issuesMap map[string][][]services.ImpactPathNode) {
-	for dependencyName := range components {
-		updatedComponent := services.Component{
-			FixedVersions: components[dependencyName].FixedVersions,
-			ImpactPaths:   issuesMap[dependencyName],
-			Cpes:          components[dependencyName].Cpes,
-		}
-		components[dependencyName] = updatedComponent
-	}
-}
-
-func setPathsForIssues(dependency *xrayUtils.GraphNode, issuesImpactPathsMap map[string][][]services.ImpactPathNode, pathFromRoot []services.ImpactPathNode) {
-	pathFromRoot = append(pathFromRoot, services.ImpactPathNode{ComponentId: dependency.Id})
-	if _, exists := issuesImpactPathsMap[dependency.Id]; exists {
-		// Create a copy of pathFromRoot to avoid modifying the original slice
-		pathCopy := make([]services.ImpactPathNode, len(pathFromRoot))
-		copy(pathCopy, pathFromRoot)
-		issuesImpactPathsMap[dependency.Id] = append(issuesImpactPathsMap[dependency.Id], pathCopy)
-	}
-	for _, depChild := range dependency.Nodes {
-		setPathsForIssues(depChild, issuesImpactPathsMap, pathFromRoot)
-	}
 }
 
 func GetMsgToUserForCurationBlock(isCurationCmd bool, tech techutils.Technology, cmdOutput string) (msgToUser string) {
